refactor(VtmnDropdown): create directive folder for clickOutside
