+++ packages/sources/svelte/src/directives/clickOutside.js
